Settings refactored. MainWindow with less slots. Autoselect usb device if only one.
Settings:
* Create settings object in main.cpp already
* Reduce code in settings.h/cpp by using c++11 class field initalisations
* Remove unnessesary color class
* Use a permanent QSettings object and introduce a setFilename method.

Bigger scope:
MainWindow was responsible to create all kind of important objects like the
DataAnalyser, the DSOControl and also the settings object. Finally it is
only responsible for assembling the main graphical interface window and
all DSO related objects are created in main. This way the current GUI can
be more easily replaced with a modern or even headless variant.