A little bit of refactoring for ChannelWidgets.
Make ChannelWidgets look more like a composite widget and layout
properly. Previously it was a widget which was never added to any layout.
Therefore it had default position (0,0) and size (100, 30) inside
it's parrent (DsoWidget). Whenever a trigger/level slider was in
the left top corner, it was getting under invisible ChannelWidgets.
It this case it was impossible to get it out from there since all
the mouse events were directed to the ChannelWidgets.