A couple of low-level fixes.
This is probably my last commit on the openhantek2 branch.
I've got it to the stage when I can see the reference signal with
 - correct offset
 - correct period
 - correct amplitude
 - correct voltage measurment label at the bottom
 - it is possible to change the sample rate and the timebase is calculated correctly

Things which are still broken:
 - software gain control is broken (and turned off in this commit)
 - with the hardware gain control, the gain dropdown shows Vols per full
   scale instead of Volts per 1 div
 - hardware gain is not initialized when application is first started
   (need to select some other gain to initialize it)
 - hardware offset is not initialized when changing the gain level
   (need to shift the slider a bit to initialize it)
 - trigger offset and position are broken
 - axes lines are still randomly broken
 - only 10kS mode currently works
 - it is impossible to change the timebase directly using the spinner

Overall it looks like the codebase is in a pretty broken state.
At this point it feels easier to start over with something working and
incrementally apply refactoring on top of it while making sure that
none of the existing logic has changed.